refactor: renaming to avoid sonarqube issues
