+++ AdSecGHTests/Helpers/Extensions/PointsTests.cs
@@ -1,25 +1,24 @@
-﻿
+﻿using System.Collections.Generic;
+
+using AdSecGH.Parameters;
+
 using Grasshopper.Kernel;
 
-using Oasys.AdSec.Materials;
-
-using OasysUnits;
-using OasysUnits.Units;
+using Rhino.Geometry;
 
 using TestGrasshopperObjects.Extensions;
 
 using Xunit;
 
 namespace AdSecGHTests.Helpers.Extensions {
-
   [Collection("GrasshopperFixture collection")]
-  public class IConcreteCrackCalculationParametersTests {
-    private IConcreteCrackCalculationParametersTestComponent _component;
-    private readonly string _failToRetrieveDataWarning = "failed";
+  public class PointsTests {
     private readonly string _convertDataError = "convert";
+    private readonly string _failToRetrieveDataWarning = "failed";
+    private PointsTestComponent _component;
 
-    public IConcreteCrackCalculationParametersTests() {
-      _component = new IConcreteCrackCalculationParametersTestComponent();
+    public PointsTests() {
+      _component = new PointsTestComponent();
     }
 
     [Fact]
@@ -55,7 +54,7 @@ namespace AdSecGHTests.Helpers.Extensions {
     [Fact]
     public void ReturnsErrorWhenDataIncorrectInputIsOptional() {
       _component.Optional = true;
-      ComponentTestHelper.SetInput(_component, string.Empty);
+      ComponentTestHelper.SetInput(_component, new List<string>());
 
       object result = ComponentTestHelper.GetOutput(_component);
       Assert.Null(result);
@@ -71,7 +70,7 @@ namespace AdSecGHTests.Helpers.Extensions {
     [Fact]
     public void ReturnsErrorWhenDataIncorrectInputIsNonOptional() {
       _component.Optional = false;
-      ComponentTestHelper.SetInput(_component, string.Empty);
+      ComponentTestHelper.SetInput(_component, new List<string>());
 
       object result = ComponentTestHelper.GetOutput(_component);
       Assert.Null(result);
@@ -80,17 +79,14 @@ namespace AdSecGHTests.Helpers.Extensions {
 
       Assert.Single(runtimeMessages);
       Assert.Contains(runtimeMessages, item => item.Contains(_convertDataError));
+      Assert.Contains(runtimeMessages, item => item.Contains("item 0"));
       Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Warning));
       Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Remark));
     }
 
     [Fact]
-    public void ReturnsIConcreteCrackCalculationParametersWhenDataCorrect() {
-      var pressure = new Pressure(-0.5, PressureUnit.Bar);
-      var pressure2 = new Pressure(1, PressureUnit.Bar);
-      var input = IConcreteCrackCalculationParameters.Create(pressure2, pressure, pressure2);
-
-      ComponentTestHelper.SetInput(_component, input);
+    public void ReturnsIPointWhenDataCorrect() {
+      ComponentTestHelper.SetInput(_component, new AdSecPointGoo(new Point3d()));
 
       object result = ComponentTestHelper.GetOutput(_component);
       Assert.NotNull(result);
@@ -98,6 +94,39 @@ namespace AdSecGHTests.Helpers.Extensions {
       Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Error));
       Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Warning));
       Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Remark));
+    }
+
+    [Fact]
+    public void ReturnsIPointWhenDataCorrectButShowRemarkForSinglePoint() {
+      ComponentTestHelper.SetInput(_component, new Point3d());
+
+      object result = ComponentTestHelper.GetOutput(_component);
+      Assert.NotNull(result);
+
+      var runtimeMessages = _component.RuntimeMessages(GH_RuntimeMessageLevel.Remark);
+
+      Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Error));
+      Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Warning));
+      Assert.Single(runtimeMessages);
+      Assert.Contains(runtimeMessages, item => item.Contains("converted to local point."));
+    }
+
+    [Fact]
+    public void ReturnsIPointWhenDataCorrectButShowRemarkForManyPoints() {
+      ComponentTestHelper.SetInput(_component, new List<object>() {
+        new Point3d(1, 1, 1),
+        new Point3d(2, 2, 2),
+      });
+
+      object result = ComponentTestHelper.GetOutput(_component);
+      Assert.NotNull(result);
+
+      var runtimeMessages = _component.RuntimeMessages(GH_RuntimeMessageLevel.Remark);
+
+      Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Error));
+      Assert.Empty(_component.RuntimeMessages(GH_RuntimeMessageLevel.Warning));
+      Assert.Single(runtimeMessages);
+      Assert.Contains(runtimeMessages, item => item.Contains("List of Points"));
     }
   }
 }
