+++ TestGrasshopperObjects/Extensions/PointsTestComponent.cs
@@ -8,14 +8,12 @@ using OasysGH;
 using OasysGH.Components;
 
 namespace TestGrasshopperObjects.Extensions {
-#pragma warning disable S101 // Types should be named in PascalCase
-  public class ILayersTestComponent : GH_OasysComponent {
-#pragma warning restore S101 // Types should be named in PascalCase
+  public class PointsTestComponent : GH_OasysComponent {
     public bool Optional { get; set; }
-    public ILayersTestComponent() : base("t0", "t1", "t2", "t3", "t4") { }
     public override Guid ComponentGuid => Guid.NewGuid();
 
     public override OasysPluginInfo PluginInfo => null;
+    public PointsTestComponent() : base("t0", "t1", "t2", "t3", "t4") { }
 
     protected override void RegisterInputParams(GH_InputParamManager pManager) {
       pManager.AddGenericParameter("test Input", "i", "input", GH_ParamAccess.list);
@@ -26,7 +24,8 @@ namespace TestGrasshopperObjects.Extensions {
     }
 
     protected override void SolveInstance(IGH_DataAccess DA) {
-      DA.SetData(0, this.GetILayers(DA, 0, Optional));
+      var result = this.GetIPoints(DA, 0, Optional);
+      DA.SetData(0, result);
     }
   }
 }
